+++ patches/src/main/kotlin/app/revanced/patches/spotify/misc/UnlockPremiumPatch.kt
@@ -2,11 +2,18 @@ package app.revanced.patches.spotify.misc
 
 import app.revanced.patcher.extensions.InstructionExtensions.addInstruction
 import app.revanced.patcher.extensions.InstructionExtensions.getInstruction
+import app.revanced.patcher.extensions.InstructionExtensions.removeInstruction
 import app.revanced.patcher.extensions.InstructionExtensions.replaceInstruction
+import app.revanced.patcher.fingerprint
 import app.revanced.patcher.patch.bytecodePatch
 import app.revanced.patches.spotify.misc.extension.sharedExtensionPatch
+import app.revanced.util.*
 import com.android.tools.smali.dexlib2.AccessFlags
+import com.android.tools.smali.dexlib2.Opcode
 import com.android.tools.smali.dexlib2.iface.instruction.OneRegisterInstruction
+import com.android.tools.smali.dexlib2.iface.instruction.TwoRegisterInstruction
+import com.android.tools.smali.dexlib2.iface.reference.FieldReference
+import com.android.tools.smali.dexlib2.iface.reference.MethodReference
 
 private const val EXTENSION_CLASS_DESCRIPTOR = "Lapp/revanced/extension/spotify/misc/UnlockPremiumPatch;"
 
@@ -27,23 +34,75 @@ val unlockPremiumPatch = bytecodePatch(
         }
 
         // Override the attributes map in the getter method.
-        val attributesMapRegister = 0
-        val instantiateUnmodifiableMapIndex = 1
-        productStateProtoFingerprint.method.addInstruction(
-            instantiateUnmodifiableMapIndex,
-            "invoke-static { v$attributesMapRegister }," +
-                "$EXTENSION_CLASS_DESCRIPTOR->overrideAttribute(Ljava/util/Map;)V",
-        )
+        with(productStateProtoFingerprint.method) {
+            val getAttributesMapIndex = indexOfFirstInstructionOrThrow(Opcode.IGET_OBJECT)
+            val attributesMapRegister = getInstruction<TwoRegisterInstruction>(getAttributesMapIndex).registerA
+
+            addInstruction(
+                getAttributesMapIndex + 1,
+                "invoke-static { v$attributesMapRegister }, " +
+                        "$EXTENSION_CLASS_DESCRIPTOR->overrideAttribute(Ljava/util/Map;)V"
+            )
+        }
 
         // Add the query parameter trackRows to show popular tracks in the artist page.
-        val addQueryParameterIndex = buildQueryParametersFingerprint.stringMatches.first().index - 1
-        buildQueryParametersFingerprint.method.replaceInstruction(addQueryParameterIndex, "nop")
+        with(buildQueryParametersFingerprint) {
+            val addQueryParameterConditionIndex = method.indexOfFirstInstructionReversedOrThrow(
+                stringMatches.first().index, Opcode.IF_EQZ
+            )
+            method.replaceInstruction(addQueryParameterConditionIndex, "nop")
+        }
 
         // Disable the "Spotify Premium" upsell experiment in context menus.
         with(contextMenuExperimentsFingerprint) {
-            val moveIsEnabledIndex = stringMatches.first().index + 2
+            val moveIsEnabledIndex = method.indexOfFirstInstructionOrThrow(
+                stringMatches.first().index, Opcode.MOVE_RESULT
+            )
             val isUpsellEnabledRegister = method.getInstruction<OneRegisterInstruction>(moveIsEnabledIndex).registerA
             method.replaceInstruction(moveIsEnabledIndex, "const/4 v$isUpsellEnabledRegister, 0")
+        }
+
+        // Make featureTypeCase_ accessible so we can check the home section type in the extension.
+        homeSectionFingerprint.classDef.fields.first { it.name == "featureTypeCase_" }.apply {
+            accessFlags = accessFlags.or(AccessFlags.PUBLIC.value).and(AccessFlags.PRIVATE.value.inv())
+        }
+
+        val protobufListClassName = with(protobufListsFingerprint.originalMethod) {
+            val emptyProtobufListGetIndex = indexOfFirstInstructionOrThrow(Opcode.SGET_OBJECT)
+            getInstruction(emptyProtobufListGetIndex).getReference<FieldReference>()!!.definingClass
+        }
+
+        val protobufListRemoveFingerprint by fingerprint {
+            custom { method, classDef ->
+                method.name == "remove" && classDef.type == protobufListClassName
+            }
+        }
+
+        // Need to allow mutation of the list so the home ads sections can be removed.
+        // Protobuffer list has an 'isMutable' boolean parameter that sets the mutability.
+        // Forcing that always on breaks unrelated code in strange ways.
+        // Instead, remove the method call that checks if the list is unmodifiable.
+        with(protobufListRemoveFingerprint.method) {
+            val invokeThrowUnmodifiableIndex = indexOfFirstInstructionOrThrow {
+                val reference = getReference<MethodReference>()
+                opcode == Opcode.INVOKE_VIRTUAL &&
+                        reference?.returnType == "V" && reference.parameterTypes.isEmpty()
+            }
+
+            // Remove the method call that throws an exception if the list is not mutable.
+            removeInstruction(invokeThrowUnmodifiableIndex)
+        }
+
+        // Remove ads sections from home.
+        with(homeStructureFingerprint.method) {
+            val getSectionsIndex = indexOfFirstInstructionOrThrow(Opcode.IGET_OBJECT)
+            val sectionsRegister = getInstruction<TwoRegisterInstruction>(getSectionsIndex).registerA
+
+            addInstruction(
+                getSectionsIndex + 1,
+                "invoke-static { v$sectionsRegister }, " +
+                        "$EXTENSION_CLASS_DESCRIPTOR->removeHomeSections(Ljava/util/List;)V"
+            )
         }
     }
 }